requirements.txt
requests
jsonschema
azure-communication-email